Update stop time in widget if stop message received
